Moved files to containers
