+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/home/mahr/Desktop/section-6/redux/bookstore/src/index.js":"1","/home/mahr/Desktop/section-6/redux/bookstore/src/components/App.js":"2","/home/mahr/Desktop/section-6/redux/bookstore/src/components/Book.js":"3","/home/mahr/Desktop/section-6/redux/bookstore/src/reducers/index.js":"4","/home/mahr/Desktop/section-6/redux/bookstore/src/reducers/books.js":"5","/home/mahr/Desktop/section-6/redux/bookstore/src/actions/booksActionTypes.js":"6","/home/mahr/Desktop/section-6/redux/bookstore/src/components/BooksForm.js":"7","/home/mahr/Desktop/section-6/redux/bookstore/src/components/BooksList.js":"8"},{"size":355,"mtime":1609949966707,"results":"9","hashOfConfig":"10"},{"size":215,"mtime":1609949861898,"results":"11","hashOfConfig":"10"},{"size":499,"mtime":1609941460632,"results":"12","hashOfConfig":"10"},{"size":124,"mtime":1609939802826,"results":"13","hashOfConfig":"10"},{"size":471,"mtime":1609939748421,"results":"14","hashOfConfig":"10"},{"size":84,"mtime":1609921550085,"results":"15","hashOfConfig":"10"},{"size":542,"mtime":1609941811316,"results":"16","hashOfConfig":"10"},{"size":876,"mtime":1609949861890,"results":"17","hashOfConfig":"10"},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},"zd7bu1",{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"31"},{"filePath":"32","messages":"33","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"34","messages":"35","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},"/home/mahr/Desktop/section-6/redux/bookstore/src/index.js",[],["36","37","38","39","40","41"],"/home/mahr/Desktop/section-6/redux/bookstore/src/components/App.js",[],"/home/mahr/Desktop/section-6/redux/bookstore/src/components/Book.js",[],"/home/mahr/Desktop/section-6/redux/bookstore/src/reducers/index.js",[],"/home/mahr/Desktop/section-6/redux/bookstore/src/reducers/books.js",[],"/home/mahr/Desktop/section-6/redux/bookstore/src/actions/booksActionTypes.js",[],["42","43","44","45","46","47"],"/home/mahr/Desktop/section-6/redux/bookstore/src/components/BooksForm.js",[],"/home/mahr/Desktop/section-6/redux/bookstore/src/components/BooksList.js",[],{"ruleId":"48","replacedBy":"49"},{"ruleId":"50","replacedBy":"51"},{"ruleId":"52","replacedBy":"53"},{"ruleId":"54","replacedBy":"55"},{"ruleId":"56","replacedBy":"57"},{"ruleId":"58","replacedBy":"59"},{"ruleId":"48","replacedBy":"60"},{"ruleId":"50","replacedBy":"61"},{"ruleId":"52","replacedBy":"62"},{"ruleId":"54","replacedBy":"63"},{"ruleId":"56","replacedBy":"64"},{"ruleId":"58","replacedBy":"65"},"lines-around-directive",["66"],"no-spaced-func",["67"],"global-require",[],"no-buffer-constructor",[],"no-new-require",[],"no-path-concat",[],["66"],["67"],[],[],[],[],"padding-line-between-statements","func-call-spacing"]
+[{"/home/mahr/Desktop/section-6/redux/bookstore/src/index.js":"1","/home/mahr/Desktop/section-6/redux/bookstore/src/components/App.js":"2","/home/mahr/Desktop/section-6/redux/bookstore/src/components/Book.js":"3","/home/mahr/Desktop/section-6/redux/bookstore/src/reducers/index.js":"4","/home/mahr/Desktop/section-6/redux/bookstore/src/reducers/books.js":"5","/home/mahr/Desktop/section-6/redux/bookstore/src/actions/booksActionTypes.js":"6","/home/mahr/Desktop/section-6/redux/bookstore/src/containers/BooksList.js":"7","/home/mahr/Desktop/section-6/redux/bookstore/src/containers/BooksForm.js":"8"},{"size":355,"mtime":1609949966707,"results":"9","hashOfConfig":"10"},{"size":239,"mtime":1609954778786,"results":"11","hashOfConfig":"10"},{"size":499,"mtime":1609941460632,"results":"12","hashOfConfig":"10"},{"size":124,"mtime":1609939802826,"results":"13","hashOfConfig":"10"},{"size":471,"mtime":1609939748421,"results":"14","hashOfConfig":"10"},{"size":84,"mtime":1609921550085,"results":"15","hashOfConfig":"10"},{"size":888,"mtime":1609954768766,"results":"16","hashOfConfig":"10"},{"size":542,"mtime":1609941811316,"results":"17","hashOfConfig":"10"},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},"zd7bu1",{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"31"},{"filePath":"32","messages":"33","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"34","messages":"35","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/home/mahr/Desktop/section-6/redux/bookstore/src/index.js",[],["36","37","38","39","40","41"],"/home/mahr/Desktop/section-6/redux/bookstore/src/components/App.js",[],"/home/mahr/Desktop/section-6/redux/bookstore/src/components/Book.js",[],"/home/mahr/Desktop/section-6/redux/bookstore/src/reducers/index.js",[],"/home/mahr/Desktop/section-6/redux/bookstore/src/reducers/books.js",[],"/home/mahr/Desktop/section-6/redux/bookstore/src/actions/booksActionTypes.js",[],["42","43","44","45","46","47"],"/home/mahr/Desktop/section-6/redux/bookstore/src/containers/BooksList.js",[],"/home/mahr/Desktop/section-6/redux/bookstore/src/containers/BooksForm.js",[],{"ruleId":"48","replacedBy":"49"},{"ruleId":"50","replacedBy":"51"},{"ruleId":"52","replacedBy":"53"},{"ruleId":"54","replacedBy":"55"},{"ruleId":"56","replacedBy":"57"},{"ruleId":"58","replacedBy":"59"},{"ruleId":"48","replacedBy":"60"},{"ruleId":"50","replacedBy":"61"},{"ruleId":"52","replacedBy":"62"},{"ruleId":"54","replacedBy":"63"},{"ruleId":"56","replacedBy":"64"},{"ruleId":"58","replacedBy":"65"},"lines-around-directive",["66"],"no-spaced-func",["67"],"global-require",[],"no-buffer-constructor",[],"no-new-require",[],"no-path-concat",[],["66"],["67"],[],[],[],[],"padding-line-between-statements","func-call-spacing"]
